further restructuring, still compiles, pages directory subdivided for convenience
